Remove Tensorflow usage in xprof/convert
The most notable changed files are:
convert/op_profile_builder.cc
plugin/tensorboard_plugin_profile/protobuf/tensorflow_datatypes.proto
utils/tensorflow_utils.cc
utils/tensorflow_utils.h

Notably, this relies on a much simpler implementation of top_n that should be sufficient for our usecase. TF's top_n is designed to be most efficient for when elements are pushed in one at a time, whereas in this case we already have the full set beforehand.

This also copies out logic relating to identifying TF datatypes. This is used specifically for the Memory Profile page, for translating a proto into a string to inform the user of which datatype the op used. This type shouldn't change often, and it doesn't feel worth it to pull in Tensorflow solely for an enum -> string conversion.

PiperOrigin-RevId: 748822688