Remove Tensorflow Dependency from XProf
This removes the circular dependency created during this migration. As a result, we need to import both xla and tsl manually. We keep them synced at head, which mimics the same behavior we had with Tensorflow.

There is one feature that we currently depend on Tensorflow for, tf_op cost analysis. For now, this feature will be disabled.

PiperOrigin-RevId: 748822690