Minimize Tensorboard dependency usage
Currently, builds end up at least partially using Tensorflow's code paths as a result of Tensorboard's Python dependencies. With the standalone profiler, we are able to build and run without using Tensorboard in Python, so we can cut it out.

The Javascript build rules are more complicated, and don't seem worth removing at the moment.

PiperOrigin-RevId: 772241874