Add source stack trace source code view to Trace Viewer page.
We use the new stack trace page component to show snippets of source code around the stack trace in a new page. This is particularly useful for Trace Viewer, since it is not written in Angular, but `StackTraceSnippet` is written in Angular. This component provides a bridge between the two.

The implementation is quite different from all the previous changes that add the source code viewer to other tools. This is because Trace Viewer is not an Angular component and a lot needed to be done to bride the gap.

In particular `window.sessionStorage` is used twice. First, Angular uses it to let Trace Viewer know whether source code service is available or not. Second, Trace Viewer uses it to pass stack trace to Angular.

We did not use GET query parameters to pass stack trace for two reasons. First, in the current settings right after the source code page shows up, the `stack_trace` parameter disappears. Second, encoding a multiline `stack_trace` in the URL makes it non-readable (users can still open multiple pages and refresh them).

PiperOrigin-RevId: 773886721