Add stack trace page.
We use the new stack trace page component to show snippets of source code around the stack trace in a new page. This is particularly useful for Trace Viewer, since it is not written in Angular, but `StackTraceSnippet` is written in Angular. This component provides a bridge between the two.

PiperOrigin-RevId: 775840015