Upgrade js/ts bazel stack
This moves to the aspect_rules_js stack, upgrades rollup, and manually uses sass for compiling material, as the old rules_sass is incompatible/deprecated. The gzgz_rules_sass repo doesn't have support for using npm packages. devversion's rules_sass currently doesn't seem able to compile the @angular/material version that we are on, but it is able to handle the remainder of our scss files.

This most likely fixes the intermittent build failures that we've seen from css.

PiperOrigin-RevId: 776765803