Upgrade js/ts bazel stack
This moves to the aspect_rules_js stack, upgrades rollup, and manually uses sass (as well as upgrading to a new repo), as the old rules_sass is incompatible/deprecated. The gzgz_rules_sass repo doesn't have support for using npm packages. devversion's rules_sass currently doesn't seem able to compile the @angular/material version that we are on.

This should fix Windows builds, as well as our intermittent build failures.

PiperOrigin-RevId: 776765803