Update protobuf dependency version
Our protos get built and compiled with `gencode 6.31.1`, and protobuf does not have compatibility outside of major versions. This also resolves an issue with using a protobuf API from a newer version.

This will cause an incompatibility with Tensorflow<=2.20. This incompatibility is already technically there, but invisible. Tensorflow==2.19.X has a hard-cap on the protobuf version, preventing it from using 6.X.X. Unfortunately, XLA depends on 6.X.X protos, so this change is required.

Resolves #1846

PiperOrigin-RevId: 821721239