[XProf: trace viewer] Only draw the visible groups
This change pre-calculates the height of each group and uses binary search to determine which groups are visible within the current viewport.

Dummy elements are used to represent the height of the culled groups above and below the visible range, ensuring smooth scrolling.

Also set `WindowPadding` to 0 to handle the binary search more easily.

This significantly improves performance by avoiding drawing off-screen groups.

There is no visual change after this CL.
The change in the tests is intended, because remove the `WindowPadding` also made horizontal available width wider.

PiperOrigin-RevId: 836976396