[XProf: trace viewer] Pre-calculate visible level offsets in Timeline.
This change introduces a `visible_level_offsets_` vector to store the y-offset for each level that is visually rendered. This allows for different rendering heights for different group types, such as flame charts and counters. The `CalculateOffsets` function now populates both group and visible level offsets. The event drawing logic is updated to use these pre-calculated offsets for positioning events.

This will be used in the navigate to event feature.

There is no visual change after this CL

PiperOrigin-RevId: 845084690