[XProf: trace viewer] Add data re-fetching logic to Trace Viewer v2 timeline.
This change introduces a mechanism to request more trace event data when the user navigates or zooms in the timeline. The timeline checks if the currently visible time range is within a "preserved" area of the already loaded data. If not, it dispatches a "request_data" event to fetch a larger "fetch" area around the current viewport.

PiperOrigin-RevId: 845951775