[XProf: trace viewer] Add event listener to handle data re-fetching in Trace Viewer V2.
The Trace Viewer V2 now listens for a 'request_data' custom event (dispatched from wasm when it determined that new data should be fetched). When this event is received, it uses the previously loaded data URL, updates it with 'start' and 'end' parameters from the event detail, and fetches new trace data. The fetched data is then processed by the trace viewer module.

PiperOrigin-RevId: 845997546