[XProf: trace viewer]Add the full trace timespan in Trace Viewer v2.
This change introduces support for `fullTimespan` field in the TraceData interface and `ParsedTraceEvents`. When provided, this full timespan is used to set the `data_time_range` in the Timeline, which dictates the boundaries for panning and zooming. This allows users to pan and zoom across the entire duration of a trace, even if only a subset of events is currently loaded. The `fetched_data_time_range` still represents the time range of the events actually loaded.

There is no visual change after this CL.

PiperOrigin-RevId: 846577431